Get rid of obs-mcp/http package to simplify
